server/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
pillow
numpy
transformers
torch
hf-xet
supabase
python-dotenv
PyPDF2
python-docx
pytest
pytest-asyncio
